requirements.txt
tree-sitter==0.24.0
tree-sitter-python==0.23.0
tree-sitter-javascript==0.23.0
tree-sitter-typescript==0.23.0
networkx==3.4
anthropic==0.42.0
openai==1.60.0
pinecone-client==5.0.0
PyGithub==2.5.0
gitpython==3.1.44
ruff
semgrep
pytest
sentence-transformers
neo4j>=5.0
